Feat: se refactoriza modelos machines y dailyProduction. Se agrega modulo de metricas

+++ WebApiCanon/Features/Machines/DTOs/MachineResponseDto.cs
@@ -3,7 +3,7 @@ namespace WebApiCanon.Features.Machines.DTOs
 {
     public class MachineResponseDto
     {
-        public int MachineId { get; set; }
+        public long MachineId { get; set; }
         public required string Name { get; set; }
         public required string Code { get; set; } 
         public bool IsActive { get; set; }
